requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2==2.9.9
boto3==1.34.34
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
apscheduler==3.10.4
requests==2.31.0
pytest==7.4.3
httpx==0.25.2
pytest-asyncio==0.21.1
